Yasir-015: Implementing necessary refactores and query parser

+++ src/module/user-feature-access-permission/user-feature-access-permission.controller.ts
@@ -1,30 +1,27 @@
 import {
-  Controller,
-  Get,
-  Post,
-  Put,
-  Delete,
-  Param,
   Body,
+  Controller,
+  Delete,
+  Get,
   Inject,
-  Res,
   Logger,
+  Param,
+  Patch,
+  Req,
+  Res,
+  UseGuards,
 } from '@nestjs/common';
-import { ApiTags, ApiOperation, ApiResponse } from '@nestjs/swagger';
+import { ApiBody, ApiOperation, ApiResponse, ApiTags, ApiBearerAuth } from '@nestjs/swagger';
 import { Response } from 'express';
-import { IUserFeatureAccessPermissionService } from './user-feature-access-permission.interface';
-import {
-  CreateUserFeatureAccessPermissionDto,
-  UpdateUserFeatureAccessPermissionDto,
-} from './user-feature-access-permission.dto';
-import { ValidateBody } from '../../common/decorators/validate.decorator';
-import {
-  createUserFeatureAccessPermissionSchema,
-  updateUserFeatureAccessPermissionSchema,
-} from './user-feature-access-permission.validation';
 import { ResponseHandler } from 'src/common/utils/response-handler';
+import { ValidateBody } from '../../common/decorators/validate.decorator';
+import { UserFeatureAccessPermissionDto } from './user-feature-access-permission.dto';
+import { IUserFeatureAccessPermissionService } from './user-feature-access-permission.interface';
+import { createUserFeatureAccessPermissionSchema } from './user-feature-access-permission.validation';
+import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
 
 @ApiTags('User Feature Access Permissions')
+@ApiBearerAuth('JWT-auth')
 @Controller('/user-feature-access-permissions')
 export class UserFeatureAccessPermissionController {
   constructor(
@@ -33,19 +30,33 @@ export class UserFeatureAccessPermissionController {
     private readonly logger: Logger,
   ) {}
 
-  @Post()
+  @Patch()
   @ApiOperation({ summary: 'Create a new user feature access permission' })
   @ApiResponse({ status: 201, description: 'Permission created successfully' })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
   @ValidateBody(createUserFeatureAccessPermissionSchema)
-  async createUserFeatureAccessPermission(
-    @Body() data: CreateUserFeatureAccessPermissionDto,
+  @ApiBody({
+    type: [UserFeatureAccessPermissionDto]
+  })
+  @UseGuards(JwtAuthGuard)
+  async createOrUpdateUserFeatureAccessPermission(
+    @Req() request: Request,
+    @Body() data: UserFeatureAccessPermissionDto[],
     @Res() response: Response,
   ) {
     return ResponseHandler.handler(
       response,
       async () => {
+        const { user } = request as any;
+        if (user.role !== 'admin') {
+          return {
+            statusCode: 403,
+            message: 'You need to be admin to set permission',
+            data: null,
+          };
+        }
         const result =
-          await this.userFeatureAccessPermissionService.createUserFeatureAccessPermission(
+          await this.userFeatureAccessPermissionService.createOrUpdateUserFeatureAccessPermission(
             data,
           );
         return {
@@ -61,6 +72,7 @@ export class UserFeatureAccessPermissionController {
   @Get()
   @ApiOperation({ summary: 'Get all user feature access permissions' })
   @ApiResponse({ status: 200, description: 'Returns all permissions' })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
   async getAllUserFeatureAccessPermissions(@Res() response: Response) {
     return ResponseHandler.handler(
       response,
@@ -80,6 +92,7 @@ export class UserFeatureAccessPermissionController {
   @Get('/:id')
   @ApiOperation({ summary: 'Get a user feature access permission by ID' })
   @ApiResponse({ status: 200, description: 'Returns the permission' })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
   async getUserFeatureAccessPermissionById(
     @Param('id') id: string,
     @Res() response: Response,
@@ -101,43 +114,27 @@ export class UserFeatureAccessPermissionController {
     );
   }
 
-  @Put('/:id')
-  @ApiOperation({ summary: 'Update a user feature access permission' })
-  @ApiResponse({ status: 200, description: 'Permission updated successfully' })
-  @ValidateBody(updateUserFeatureAccessPermissionSchema)
-  async updateUserFeatureAccessPermission(
-    @Param('id') id: string,
-    @Body() data: UpdateUserFeatureAccessPermissionDto,
-    @Res() response: Response,
-  ) {
-    return ResponseHandler.handler(
-      response,
-      async () => {
-        const permission =
-          await this.userFeatureAccessPermissionService.updateUserFeatureAccessPermission(
-            id,
-            data,
-          );
-        return {
-          statusCode: 200,
-          message: 'Permission updated successfully',
-          data: permission,
-        };
-      },
-      this.logger,
-    );
-  }
-
   @Delete('/:id')
   @ApiOperation({ summary: 'Delete a user feature access permission' })
   @ApiResponse({ status: 200, description: 'Permission deleted successfully' })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
+  @UseGuards(JwtAuthGuard)
   async deleteUserFeatureAccessPermission(
+    @Req() request: Request,
     @Param('id') id: string,
     @Res() response: Response,
   ) {
     return ResponseHandler.handler(
       response,
       async () => {
+        const { user } = request as any;
+        if (user.role !== 'admin') {
+          return {
+            statusCode: 403,
+            message: 'You need to be admin to change permission',
+            data: null,
+          };
+        }
         const permission =
           await this.userFeatureAccessPermissionService.deleteUserFeatureAccessPermission(
             id,
